Fix bug
Once put into unavaiable, people should not
appear in any other list. This happened because -- uses equality to
remove elements from a given list. Since in the UI we need to keep track
of position to make SortableJS happy, this would no longer work as the
position of people in different columns would of course rarely align