Get tests green again on CI
Strangely enough this only happened when the test database was empty.
Running mix test.watch was not sufficient to catch this, and in fact
running mix test.watch seemed to mask this bug.

The problem was that the tracks we would iterate were not consistent
and when the test database was empty they would come back in an
unexpected order.