Fix crash
This was crashing anytime you assigned someone to a track. Moving from
using the track name as the unique id for the sortable.js library to
using the track primary key (which seemed like a good idea at the time)
revealed the fact that we were using the track name passed back to the
live view to facilitate putting the UI in the correct state and to
update the database.

I regret past Tim's decisions.